refactor: move mixins and utilities out of components
This is done in preperation for a new group of components relying girder
api endpoints.  These will need to share some of the mixins from other
component groups, and I want to avoid importing packages between groups.
