+++ src/bindWatchers.js
