+++ src/components/geojs/index.js
@@ -4,8 +4,6 @@ import GeojsHeatmapLayer from './GeojsHeatmapLayer';
 import GeojsMapViewport from './GeojsMapViewport';
 import GeojsTileLayer from './GeojsTileLayer';
 import GeojsWidgetLayer from './GeojsWidgetLayer';
-import bindWatchers from './bindWatchers';
-import mixins from './mixins';
 
 export {
   GeojsAnnotationLayer,
@@ -14,6 +12,4 @@ export {
   GeojsMapViewport,
   GeojsTileLayer,
   GeojsWidgetLayer,
-  bindWatchers,
-  mixins,
 };
