+++ src/mixins/geojsLayer.js
@@ -18,4 +18,4 @@ const layerMixin = {
   },
 };
 
-export { layerMixin }; // eslint-disable-line import/prefer-default-export
+export default layerMixin;
