+++ src/mixins/index.js
@@ -1,0 +1,3 @@
+import geojsLayer from './geojsLayer';
+
+export { geojsLayer }; // eslint-disable-line import/prefer-default-export
